Добавляет разделение по модулям
